@@ -1,11 +1,11 @@
 name: easy_mask
-version: 1.1.4
+version: 1.2.0
 
 description: Easy and Powerful Mask for TextField, TextFormField and Strings. Has multi mask support, reverse mask for currency and placeholder. It controls user's cursor.
 homepage: https://github.com/danilocoppi/flutter-textfield-mask
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
